+++ src/viat/_cli/commands/tracked.py
@@ -8,9 +8,10 @@ from viat.vault import autoload_vault
 @viat.command()
 @click.option('--json', '-j', 'output_json', is_flag=True, help='Print the list in JSON format.')
 @click.option('--no-data', is_flag=True, help='Print only those paths without any recorded attributes.')
-def tracked(output_json: bool, no_data: bool) -> None:
+@click.pass_context
+def tracked(ctx: click.Context, output_json: bool, no_data: bool) -> None:
     """Print out the tracked file paths."""
-    vault = autoload_vault()
+    vault = autoload_vault(ctx.obj.vault_config)
 
     if no_data:
         with vault.storage as conn: