+++ src/viat/_cli/state.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+from viat.vault import ViatVaultStaticConfig
+
+
+@dataclass
+class ViatCliSharedState:
+    vault_config: ViatVaultStaticConfig