+++ src/viat/_config/storage.py
@@ -1,4 +1,5 @@
 from viat._config.loader import ConfigLoader
+from viat._vault_config import ViatVaultStaticConfig
 from viat.exceptions import ViatConfigError
 from viat.protocols import ViatAttributeStorage
 from viat.providers.storage import (
@@ -10,35 +11,43 @@ from viat.providers.storage import (
 from viat.support.path_resolver import ViatPathResolver
 
 
-def load_storage_from_config(resolver: ViatPathResolver, loader: ConfigLoader) -> ViatAttributeStorage:
+def load_storage_from_config(resolver: ViatPathResolver, static_config: ViatVaultStaticConfig, loader: ConfigLoader) -> ViatAttributeStorage:
     match loader.get_str('storage', 'provider'):
         case 'toml' | None:
-            return load_toml_storage_from_config(resolver, loader)
+            return load_toml_storage_from_config(resolver, static_config, loader)
 
         case 'json':
-            return load_json_storage_from_config(resolver, loader)
+            return load_json_storage_from_config(resolver, static_config, loader)
 
         case _:
             raise ViatConfigError('The storage.provider option must be either "toml" or "json"')
 
 
-def load_toml_storage_from_config(resolver: ViatPathResolver, loader: ConfigLoader) -> TomlAttributeStorage:
+def load_toml_storage_from_config(resolver: ViatPathResolver, static_config: ViatVaultStaticConfig, loader: ConfigLoader) -> TomlAttributeStorage:
     storage_path = loader.get_path('storage', 'toml', 'storage_path', root=resolver.get_viat(), default=resolver.get_viat() / 'storage.toml')
-    schema_path = loader.get_path('storage', 'toml', 'json_schema_path', root=resolver.get_viat())
 
-    if schema_path is None and resolver.get_viat().joinpath('schema.json').exists():
-        schema_path = resolver.get_viat() / 'schema.json'
+    if static_config.skip_validation:
+        schema_path = None
+    else:
+        schema_path = loader.get_path('storage', 'toml', 'json_schema_path', root=resolver.get_viat())
+
+        if schema_path is None and resolver.get_viat().joinpath('schema.json').exists():
+            schema_path = resolver.get_viat() / 'schema.json'
 
     storage_config = TomlAttributeStorageConfig(storage_path, schema_path)
     return TomlAttributeStorage(storage_config)
 
 
-def load_json_storage_from_config(resolver: ViatPathResolver, loader: ConfigLoader) -> JsonAttributeStorage:
+def load_json_storage_from_config(resolver: ViatPathResolver, static_config: ViatVaultStaticConfig, loader: ConfigLoader) -> JsonAttributeStorage:
     storage_path = loader.get_path('storage', 'json', 'storage_path', root=resolver.get_viat(), default=resolver.get_viat() / 'storage.json')
-    schema_path = loader.get_path('storage', 'json', 'json_schema_path', root=resolver.get_viat())
 
-    if schema_path is None and resolver.get_viat().joinpath('schema.json').exists():
-        schema_path = resolver.get_viat() / 'schema.json'
+    if static_config.skip_validation:
+        schema_path = None
+    else:
+        schema_path = loader.get_path('storage', 'json', 'json_schema_path', root=resolver.get_viat())
+
+        if schema_path is None and resolver.get_viat().joinpath('schema.json').exists():
+            schema_path = resolver.get_viat() / 'schema.json'
 
     indent = loader.get_nested('storage', 'json', 'indent')