+++ src/viat/_vault_config.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class ViatVaultStaticConfig:
+    """Static (i.e. known before initialization) configuration for the vault."""
+
+    skip_validation: bool = False
+    """Skip schema validation."""
+